+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/mp4/MetadataUtil.java
@@ -284,25 +284,22 @@ import com.google.android.exoplayer2.util.ParsableByteArray;
 
   private MetadataUtil() {}
 
-  /**
-   * Returns a {@link Format} that is the same as the input format but includes information from the
-   * specified sources of metadata.
-   */
-  public static Format getFormatWithMetadata(
+  /** Updates a {@link Format.Builder} to include metadata from the provided sources. */
+  public static void setFormatMetadata(
       int trackType,
-      Format format,
       @Nullable Metadata udtaMetadata,
       @Nullable Metadata mdtaMetadata,
-      GaplessInfoHolder gaplessInfoHolder) {
+      GaplessInfoHolder gaplessInfoHolder,
+      Format.Builder formatBuilder) {
     if (trackType == C.TRACK_TYPE_AUDIO) {
       if (gaplessInfoHolder.hasGaplessInfo()) {
-        format =
-            format.copyWithGaplessInfo(
-                gaplessInfoHolder.encoderDelay, gaplessInfoHolder.encoderPadding);
+        formatBuilder
+            .setEncoderDelay(gaplessInfoHolder.encoderDelay)
+            .setEncoderPadding(gaplessInfoHolder.encoderPadding);
       }
       // We assume all udta metadata is associated with the audio track.
       if (udtaMetadata != null) {
-        format = format.copyWithMetadata(udtaMetadata);
+        formatBuilder.setMetadata(udtaMetadata);
       }
     } else if (trackType == C.TRACK_TYPE_VIDEO && mdtaMetadata != null) {
       // Populate only metadata keys that are known to be specific to video.
@@ -311,12 +308,11 @@ import com.google.android.exoplayer2.util.ParsableByteArray;
         if (entry instanceof MdtaMetadataEntry) {
           MdtaMetadataEntry mdtaMetadataEntry = (MdtaMetadataEntry) entry;
           if (MDTA_KEY_ANDROID_CAPTURE_FPS.equals(mdtaMetadataEntry.key)) {
-            format = format.copyWithMetadata(new Metadata(mdtaMetadataEntry));
+            formatBuilder.setMetadata(new Metadata(mdtaMetadataEntry));
           }
         }
       }
     }
-    return format;
   }
 
   /**